move Header component into compponents folder
